+++ .github/dependabot.yml
@@ -21,3 +21,11 @@ updates:
       - dependency-name: "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
       - dependency-name: "org.jetbrains.kotlin.plugin.spring"
       - dependency-name: "org.jetbrains.skiko:skiko-js-wasm-runtime"
+      - dependency-name: "org.jetbrains.compose.runtime:runtime"
+      - dependency-name: "org.jetbrains.compose.ui:ui"
+      - dependency-name: "org.jetbrains.compose.animation:animation"
+      - dependency-name: "org.jetbrains.compose.animation:animation-graphics"
+      - dependency-name: "org.jetbrains.compose.foundation:foundation"
+      - dependency-name: "org.jetbrains.compose.material:material"
+      - dependency-name: "org.jetbrains.compose.material3:material3"
+      - dependency-name: "org.jetbrains.compose.components:components-resources"
