--- src/main/kotlin/com/compiler/server/model/bean/Dependencies.kt
@@ -1,6 +1,0 @@
-package com.compiler.server.model.bean
-
-data class Dependencies(
-    val composeWasm: String,
-    val staticUrl: String,
-)
